requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.6.1
pyyaml==6.0.1
pytest==7.4.0
matplotlib
seaborn
dash
plotly
sklearn.preprocessing
